[Storage server] Split unit tests into separate files.
