+++ state-sync/storage-service/server/src/tests/mod.rs
@@ -1,0 +1,19 @@
+// Copyright © Aptos Foundation
+// SPDX-License-Identifier: Apache-2.0
+
+mod cache;
+mod epoch_ending;
+mod mock;
+mod new_transaction_outputs;
+mod new_transactions;
+mod new_transactions_or_outputs;
+mod number_of_states;
+mod protocol_version;
+mod request_moderator;
+mod state_values;
+mod storage_summary;
+mod subscription;
+mod transaction_outputs;
+mod transactions;
+mod transactions_or_outputs;
+mod utils;
